requirements.txt
pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
umap-learn
jupyter
notebook
streamlit
plotly
